pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["동욱 <frlaehdd@naver.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "~4.0"
djangorestframework = "^3.13.1"
strawberry-graphql = {extras = ["debug-server"], version = "^0.130.4"}
strawberry-graphql-django = "^0.5.1"
PyJWT = "^2.4.0"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
django-environ = "^0.10.0"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
